requirements.txt
streamlit
yfinance
finnhub-python
pandas
vaderSentiment
joblib
scikit-learn==1.5.1